d_a_npc_aj1: talk_1 95.89->99.82 — int partsNotMov local (u8 local + u8-returning chk_parts_notMov elides the caller mask; int local forces u8->int clrlwi at assignment matching retail); fresh .o postdates

+++ src/d/actor/d_a_npc_aj1.cpp
@@ -1258,7 +1258,7 @@ s32 daNpc_Aj1_c::wait_1() {
 /* 00002BC8-00002D08       .text talk_1__11daNpc_Aj1_cFv */
         /* Nonmatching */
 int daNpc_Aj1_c::talk_1() {
-    u8 partsNotMov = chk_parts_notMov();
+    int partsNotMov = chk_parts_notMov();
     talk(1);
     if (mpCurrMsg == NULL) {
         return partsNotMov;
